// pharmacy-ui/src/app/services/models/order-admin-update-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface OrderAdminUpdateRequest {
  action?: 'APPROVE' | 'PREPARE' | 'DISPATCH' | 'DELIVER' | 'CANCEL';
  carrier?: string;
  departureTime?: string;
  destination?: string;
  destinationLat?: number;
  destinationLng?: number;
  expectedArrival?: string;
  notifyCustomer?: boolean;
  origin?: string;
  originLat?: number;
  originLng?: number;
  service?: string;
  shippingDate?: string;
  totalTime?: string;
}