// pharmacy-ui/src/app/services/models/shipment-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ShipmentTimelineResponse } from '../models/shipment-timeline-response';
export interface ShipmentResponse {
  carrier?: string;
  departureTime?: string;
  destination?: string;
  destinationLat?: number;
  destinationLng?: number;
  expectedArrival?: string;
  id?: string;
  orderRef?: string;
  origin?: string;
  originLat?: number;
  originLng?: number;
  progressStep?: number;
  service?: string;
  shippingDate?: string;
  status?: 'IN_PROGRESS' | 'ARRIVED' | 'CANCELED' | 'DRAFT' | 'DELAYED';
  timeline?: Array<ShipmentTimelineResponse>;
  totalTime?: string;
}